app/main.py
"""
Main FastAPI application module for the 100x-orchestrator
"""
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel

app = FastAPI(
    title="100x-orchestrator",
    description="An orchestration system for managing AI coding agents",
    version="0.1.0"
)

class AgentConfig(BaseModel):
    """Configuration model for AI agents"""
    agent_id: str
    repository_url: str | None = None
    local_path: str | None = None
    resource_allocation: dict | None = None

@app.get("/")
async def root():
    """Root endpoint returning service status"""
    return {"status": "active", "service": "100x-orchestrator"}

@app.post("/agents/start")
async def start_agent(config: AgentConfig):
    """Start a new AI coding agent"""
    # TODO: Implement agent initialization logic
    return {"status": "initialized", "agent_id": config.agent_id}

@app.post("/agents/{agent_id}/stop")
async def stop_agent(agent_id: str):
    """Stop an active AI coding agent"""
    # TODO: Implement agent termination logic
    return {"status": "terminated", "agent_id": agent_id}

@app.get("/agents/{agent_id}/status")
async def get_agent_status(agent_id: str):
    """Get the current status of an AI coding agent"""
    # TODO: Implement status checking logic
    return {"agent_id": agent_id, "status": "running"}

@app.put("/agents/{agent_id}/config")
async def update_agent_config(agent_id: str, config: AgentConfig):
    """Update the configuration of an active agent"""
    # TODO: Implement configuration update logic
    return {"status": "updated", "agent_id": agent_id}